% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace-util.R
\name{tu_apply}
\alias{tu_apply}
\title{Apply the FUN function to the observation obs for each xp in xps}
\usage{
tu_apply(xps, obs, FUN)
}
\arguments{
\item{xps}{a list of xp names}

\item{obs}{the name of one observation}

\item{FUN}{the function to apply to dataframes named xp.obs}
}
\value{
a dataframe having xp and value column
}
\description{
Apply the FUN function to the observation obs for each xp in xps
}
\examples{
Given xps=data.frame(xp=c('balancer','consolidator')
tu_apply(xps, 'used_cores', tu_integrate) will return
                 xp    value
1          balancer 402836.8
2      consolidator 402836.8
Note that xps can be loaded with source('data/reads.R')
}
\keyword{traceutil}

