% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace-util.R
\name{tu_integrate}
\alias{tu_integrate}
\title{Return the integrate of the given df over time
for each entity, or for all entities according to per_entity
Works  with numeric events and count-if traces.}
\usage{
tu_integrate(df, per_entity = FALSE)
}
\arguments{
\item{df}{a dataframe ['entity', 'date', 'value']}

\item{per_entity}{TRUE to return the integral per entity}
}
\value{
a dataframe ['entity', 'integral'] of just integral according to per_entity
}
\description{
Return the integrate of the given df over time
for each entity, or for all entities according to per_entity
Works  with numeric events and count-if traces.
}
\examples{
tu_integrate(balancer.used_core)
[1] 402836.8
}
\keyword{traceutil}

