% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace-util.R
\name{tu_intervals}
\alias{tu_intervals}
\title{Compute the intervals of a given df.}
\usage{
tu_intervals(df)
}
\arguments{
\item{df}{a dataframe ['entity', 'date', 'value']}
}
\value{
a dataframe ['entity', 'start_date', 'value', 'end_date', 'duration']
}
\description{
Compute the intervals of a given df.
}
\examples{
head(tu_intervals(balancer.used_cores))
                                                           entity start_date
1   root:cloud:myCloud:compute:compute_host:node-20.me:used_cores     0.0000
38  root:cloud:myCloud:compute:compute_host:node-20.me:used_cores    27.0000
63  root:cloud:myCloud:compute:compute_host:node-20.me:used_cores    72.0000
87  root:cloud:myCloud:compute:compute_host:node-20.me:used_cores   123.0000
111 root:cloud:myCloud:compute:compute_host:node-20.me:used_cores   204.0000
130 root:cloud:myCloud:compute:compute_host:node-20.me:used_cores   348.0078
    value end_date duration
1       4  27.0000  27.0000
38      5  72.0000  45.0000
63      6 123.0000  51.0000
130     7 378.0078  30.0000
87      7 204.0000  81.0000
111     8 348.0078 144.0078
}
\keyword{traceutil}

