% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace-util.R
\name{tu_plot}
\alias{tu_plot}
\title{Plot the given dataframe according to its type}
\usage{
tu_plot(df, title = deparse(substitute(df)))
}
\arguments{
\item{df}{a dataframe ['entity', 'date', 'value']}

\item{title}{the main title of the plot}
}
\value{
a plot
}
\description{
Plot the given dataframe according to its type
}
\examples{
tu_plot(balancer.vm__state, 'the state of vms in the balancer simulation')
}
\keyword{traceutil}

