% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace-util.R
\name{tu_plot_state}
\alias{tu_plot_state}
\title{Plot the state of entities as colored rectangles.}
\usage{
tu_plot_state(df, title = deparse(substitute(df)))
}
\arguments{
\item{df}{a dataframe ['entity', 'date', 'value'] where value is a state}

\item{title}{the main title of the plot}
}
\value{
a plot
}
\description{
Plot the state of entities as colored rectangles.
}
\examples{
tu_plot_state(balancer.vm__state)
}
\keyword{traceutil}

