% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace-util.R
\name{tu_read}
\alias{tu_read}
\title{Reads all the dat file in a given directory,
create one dataframe for each,
plot according to plotting,
and create xps with the list of xp.}
\usage{
tu_read(dir = ".", plotting = FALSE)
}
\arguments{
\item{dir}{the directory containing the dat files}

\item{plotting}{plot the data if TRUE}
}
\value{
the list of read dataframes
}
\description{
Reads all the dat file in a given directory,
create one dataframe for each,
plot according to plotting,
and create xps with the list of xp.
}
\examples{
tu_read('./data', TRUE)
will read all dit files inf the data directory and plot everything
}
\keyword{traceutil}

