% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace-util.R
\name{tu_valueat}
\alias{tu_valueat}
\title{Returns the value of a df at a given date.
Works  with events and count-if traces.}
\usage{
tu_valueat(df, date)
}
\arguments{
\item{df}{a dataframe having date and value columns}

\item{date}{a date in the simulation, in second}
}
\value{
a dataframe having entity and value column
}
\description{
Returns the value of a df at a given date.
Works  with events and count-if traces.
}
\examples{
tu_valueat(balancer.used_core, 300)
1  root:cloud:myCloud:compute:compute_host:node-2.me:used_cores     5
2  root:cloud:myCloud:compute:compute_host:node-1.me:used_cores     7
3  root:cloud:myCloud:compute:compute_host:node-3.me:used_cores     8
}
\keyword{traceutil}

